Rename mcpd_sdk to mcpd
* Rename mcpd_sdk to mcpd
* Remove __version__
* uv.lock files updated
