+++ examples/anyagent/pyproject.toml
@@ -8,7 +8,7 @@ requires-python = ">=3.11"
 dependencies = [
     "any-agent[all]>=0.24.0",
     "requests>=2.32.4",
-    "mcpd_sdk",
+    "mcpd",
 ]
 
 [dependency-groups]
@@ -34,7 +34,7 @@ all = [
 ]
 
 [tool.uv.sources]
-mcpd_sdk = { path = "../..", editable = true }
+mcpd = { path = "../..", editable = true }
 
 [tool.ruff]
 fix = true
