+++ examples/readme_maker/pyproject.toml
@@ -7,7 +7,7 @@ license = {text = "Apache-2.0"}
 requires-python = ">=3.11"
 dependencies = [
     "requests>=2.32.4",
-    "mcpd_sdk",
+    "mcpd",
     "pydantic-ai>=0.3.7",
 ]
 
@@ -34,7 +34,7 @@ all = [
 ]
 
 [tool.uv.sources]
-mcpd_sdk = { path = "../..", editable = true }
+mcpd = { path = "../..", editable = true }
 
 [tool.ruff]
 fix = true
