+++ src/mcpd/__init__.py
@@ -1,0 +1,6 @@
+from .mcpd_client import McpdClient, McpdError
+
+__all__ = [
+    "McpdClient",
+    "McpdError",
+]
