+++ src/mcpd/exceptions.py
